+++ src/ts/models/constants.model.ts
@@ -3,7 +3,6 @@ export interface Constants {
     ladingPageWidth: number,
     ladingPageHeight: number,
     radius: number,
-    amplification: number,
     stepDistanceXAxis: number,
     stepDistanceYAxis: number,
     canvasInternalPadding: number,